# Add an optimized token router filter