--- a/Cargo.toml
+++ b/Cargo.toml
@@ -58,6 +58,11 @@ name = "misc"
 harness = false
 test = false
 
+[[bench]]
+name = "token_router"
+harness = false
+test = false
+
 [dependencies]
 # Local
 quilkin-macros = { version = "0.9.0-dev", path = "./macros" }