--- a/benches/cluster_map.rs
+++ b/benches/cluster_map.rs
@@ -3,6 +3,8 @@ use quilkin::net::cluster::ClusterMap;
 
 mod shared;
 
+use shared::TokenKind;
+
 #[divan::bench_group(sample_count = 10)]
 mod serde {
     use super::*;
@@ -66,21 +68,21 @@ mod serde {
 
     #[divan::bench(consts = SEEDS)]
     fn serialize_proto<const S: u64>(b: Bencher) {
-        let gc = gen_cluster_map::<S>();
+        let gc = gen_cluster_map::<S>(TokenKind::None);
         b.counter(gc.total_endpoints)
             .bench(|| divan::black_box(serialize_to_protobuf(&gc.cm)));
     }
 
     #[divan::bench(consts = SEEDS)]
     fn serialize_json<const S: u64>(b: Bencher) {
-        let gc = gen_cluster_map::<S>();
+        let gc = gen_cluster_map::<S>(TokenKind::None);
         b.counter(gc.total_endpoints)
             .bench(|| divan::black_box(serialize_to_json(&gc.cm)));
     }
 
     #[divan::bench(consts = SEEDS)]
     fn deserialize_json<const S: u64>(b: Bencher) {
-        let gc = gen_cluster_map::<S>();
+        let gc = gen_cluster_map::<S>(TokenKind::None);
         let json = serialize_to_json(&gc.cm);
 
         b.with_inputs(|| json.clone())
@@ -90,7 +92,7 @@ mod serde {
 
     #[divan::bench(consts = SEEDS)]
     fn deserialize_proto<const S: u64>(b: Bencher) {
-        let gc = gen_cluster_map::<S>();
+        let gc = gen_cluster_map::<S>(TokenKind::None);
         let pv = serialize_to_protobuf(&gc.cm);
 
         b.with_inputs(|| pv.clone())
@@ -125,7 +127,7 @@ mod ops {
 
     #[divan::bench(consts = SEEDS)]
     fn iterate<const S: u64>(b: Bencher) {
-        let cm = gen_cluster_map::<S>();
+        let cm = gen_cluster_map::<S>(TokenKind::None);
 
         b.counter(cm.total_endpoints)
             .bench_local(|| divan::black_box(compute_hash::<S>(&cm)));
@@ -135,7 +137,7 @@ mod ops {
 
     #[divan::bench(consts = SEEDS)]
     fn iterate_par<const S: u64>(b: Bencher) {
-        let cm = gen_cluster_map::<S>();
+        let cm = gen_cluster_map::<S>(TokenKind::None);
 
         b.counter(cm.total_endpoints)
             .bench(|| divan::black_box(compute_hash::<S>(&cm)))